requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision
opencv-python>=4.8.0
mediapipe>=0.10.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# NLP & Embeddings
transformers>=4.30.0
sentence-transformers>=2.2.0

# OCR
easyocr>=1.7.0

# Audio Processing & Speech Recognition
openai-whisper>=20231117
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0

# Video Editing
opentimelineio>=0.15.0

# Configuration
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
tqdm>=4.65.0
pillow>=10.0.0

# Optional: For GPU acceleration
# cudatoolkit>=11.8  # Install via conda if needed
